Version 2.0.1

+++ includes/edit-lock.php
@@ -243,12 +243,19 @@ function bp_docs_cancel_edit_link() {
 
 		$doc_id = get_queried_object_id();
 
-		if ( !$doc_id )
-			return false;
+		if ( $doc_id ) {
+			$doc_permalink = bp_docs_get_doc_link( $doc_id );
 
-		$doc_permalink = bp_docs_get_doc_link( $doc_id );
-
-		$cancel_link = add_query_arg( 'bpd_action', 'cancel_edit', $doc_permalink );
+			$cancel_link = add_query_arg( 'bpd_action', 'cancel_edit', $doc_permalink );
+		} else {
+			// This is a cancel request on a new (or otherwise unknown) doc.
+			if ( $previous_page = wp_get_referer() ) {
+				$cancel_link = $previous_page;
+			} else {
+				$cancel_link = bp_docs_get_archive_link();
+			}
+			$doc_permalink = null;
+		}
 
 		return apply_filters( 'bp_docs_get_cancel_edit_link', $cancel_link, $doc_permalink );
 	}
@@ -261,10 +268,14 @@ function bp_docs_cancel_edit_link() {
  * @since 1.1
  */
 function bp_docs_remove_edit_lock() {
-	$doc_id = isset( $_POST['doc_id'] ) ? $_POST['doc_id'] : false;
+	if ( ! isset( $_POST['nonce'] ) || ! wp_verify_nonce( $_POST['nonce'], 'bp_docs_save' ) ) {
+		return;
+	}
 
-	if ( !$doc_id )
-		return false;
+	$doc_id = isset( $_POST['doc_id'] ) ? $_POST['doc_id'] : false;
+	if ( ! $doc_id || ! current_user_can( 'bp_docs_edit', $doc_id ) ) {
+		return;
+	}
 
 	delete_post_meta( $doc_id, '_bp_docs_last_pinged' );
 }
